Initial changes for solaris support

+++ gcl/h/att_ext.h
@@ -27,7 +27,7 @@ char *alloc_page();
 object alloc_object();
 void *malloc(size_t);
 void *realloc(void *,size_t);
-void * memalign(long,size_t);
+void * memalign(size_t,size_t);
 void *alloc_contblock(size_t);
 void *alloc_relblock(size_t);
 object fSallocate_contiguous_pages();
